Allow dbus processing in qt-base to enable color-picking via portal
qt-base uses dbus to query the desktop portal if color-picking is
supported, without explicitly waiting for the result.
kcolorchooser creates its QColorDialog before the response was processed
and therefore color picking via the portal is presumed to be
unavailable.

Give the eventloop the opportunity to process the event and only
afterwards create the QColorDialog.

See https://bugreports.qt.io/browse/QTBUG-120957

BUG: 479406